+++ frontend/host-app/src/components/Questions/QuestionIconButton/RestoreQuestionIconButton.js
@@ -1,22 +1,23 @@
-import React from "react";
+import useStyles from "../useStyles.js";
 import Tooltip from "@material-ui/core/Tooltip";
 import {Icon} from "@material-ui/core";
-import useStyles from "./useStyles.js";
+import React from "react";
 
-function ApproveQuestionIconButton(props) {
+function RestoreQuestionIconButton(props) {
 	const classes = useStyles();
 
 	return (
-		<Tooltip title="승인">
+		<Tooltip title="질문 되살리기">
 			<Icon
-				className={classes.approveButton}
+				className={classes.restoreButton}
 				onClick={() =>
 					props.dataHandler(props.id, props.type, "active")
 				}
 			>
-				check_circle_outline
+				restore
 			</Icon>
 		</Tooltip>
 	);
 }
-export default ApproveQuestionIconButton;
+
+export default RestoreQuestionIconButton;
